requirements.txt
# RAG API测试脚本依赖包
python-dotenv>=1.0.0
anthropic>=0.25.0